requirements.txt
streamlit
google
natsort
dotenv
reportlab
python-docx